+++ mods-src/farming/src/trough.rs
@@ -90,11 +90,11 @@ pub fn on_interact(
     Outcome::Cancel
 }
 
-fn center(pos: [i32; 3]) -> [f32; 3] {
+fn center(pos: [i32; 3]) -> [f64; 3] {
     [
-        pos[0] as f32 + 0.5,
-        pos[1] as f32 + 0.5,
-        pos[2] as f32 + 0.5,
+        pos[0] as f64 + 0.5,
+        pos[1] as f64 + 0.5,
+        pos[2] as f64 + 0.5,
     ]
 }
